# back/internal/store/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.26.0

package sqlc

import (
	"time"

	"github.com/jackc/pgx/v5/pgtype"
)

type Task struct {
	ID          int64              `json:"id"`
	UserID      int64              `json:"userId"`
	Title       string             `json:"title"`
	Description pgtype.Text        `json:"description"`
	DueDate     pgtype.Timestamptz `json:"dueDate"`
	Priority    int16              `json:"priority"`
	IsCompleted bool               `json:"isCompleted"`
	CreatedAt   time.Time          `json:"createdAt"`
	UpdatedAt   time.Time          `json:"updatedAt"`
}

type User struct {
	ID        int64     `json:"id"`
	GoogleID  string    `json:"googleId"`
	Name      string    `json:"name"`
	CreatedAt time.Time `json:"createdAt"`
	UpdatedAt time.Time `json:"updatedAt"`
}